Loosen job type (#5)
* Loosen the Job type on JobChannel events

SSE Gateway is the only repo in jenkinsci org that has a dep on this, so not making the change backward compat. The purests would not be happy with that, but I'm okay with taking some flak if by chance someone notices :)

* Fix bug in GuavaPubsubBusItemTest

* More multibranch stuff for Cliff

* Added a FAQ to the readme

* Added comment in readme about alternative bus impls and Messaging Bridge pattern.

* Change to a hpi

The repo still has a "module" suffix on the name. That will probably cause upset for some people :(

* Comment cleanups in SyncQueueListener

* reinstate deprecated version of getJob on JobChannelMessage

* Move call to shutdown hook from GuavaPubsubBus to PubsubBus

* Cleanup queue task complete publishing shutdown

* Added plugin index.jelly